celery/requirements.txt
amqp==5.3.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ffmpeg-python==0.2.0
future==1.0.0
kombu==5.4.2
libtorrent==2.0.11
prompt_toolkit==3.0.50
python-dateutil==2.9.0.post0
setuptools==75.8.0
six==1.17.0
tzdata==2025.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
